+++ CHANGELOG.md
@@ -3,11 +3,12 @@
 ## [5.1.0] - 2025-10-
 ## Added
 - Audio functions: mute(), unmute() and isMuted().
-- String args alongside PlaybackButtonLayout enum.
 - Added audio-synth as an own module.
+- Added PlaybackButtons properties singlePlayStop, playStop and playPauseStop (deprecated buttonLayout).
 
 ## Fixed
 - Documentation errors.
+
 
 ## [5.0.0] - 2025-10-05
 Major update required to enable independent browser iife instrument modules.